Move DevoloMultiLevelSwitchDeviceEntity in devolo Home Control
